knative/ML/ml-pca/requirements.txt
minio==7.1.13
numpy==1.24.2
parliament-functions==0.1.0
